feat(web): automate a single artifact from its ⋯ menu
The per-artifact entry point: on any doc you own, ⋯ → Automate opens a dialog that
creates an automation scoped to THIS artifact — its short id rides along as a ref and
the instruction is seeded to 'Keep this document current.' Same agent + trigger
(Manual / Schedule / Event) + route flow as Settings → Automations, framed for one doc.

Extracted the create form into a shared AutomationForm (settings + the artifact dialog
both use it, no duplication); the settings manager is behavior-unchanged. Gated on
canMove (owner) so only someone who administers the workspace sees it.

Also: the shared form handles its agents-query error (surfaces gate), and dropped a
genuinely-unused 'lt' import in pg.ts (biome unsafe-fix, left by the earlier refactor).

web typecheck + tests 182 + testids/tokens/mutations/frontend/surfaces + biome green.
