+++ apps/web/src/pages/settings/automations-section.tsx
@@ -1,32 +1,18 @@
 import { useQuery, useQueryClient } from "@tanstack/react-query"
 import { Zap } from "lucide-react"
-import { useMemo, useState } from "react"
-import { type Automation, type AutomationTrigger, api, type Run } from "@/api"
+import { useState } from "react"
+import { type Automation, api, type Run } from "@/api"
 import { ConfirmDialog } from "@/components/shared/confirm-dialog"
 import { EmptyState } from "@/components/shared/empty-state"
 import { SettingsGroup } from "@/components/shared/settings-group"
 import { StatusPanel } from "@/components/shared/status-panel"
 import { Badge } from "@/components/ui/badge"
 import { Button } from "@/components/ui/button"
-import {
-  Select,
-  SelectContent,
-  SelectItem,
-  SelectTrigger,
-  SelectValue,
-} from "@/components/ui/select"
-import { Tabs, TabsList, TabsTrigger } from "@/components/ui/tabs"
-import { Textarea } from "@/components/ui/textarea"
-import { agentsQuery, automationsQuery, runsQuery } from "@/lib/queries"
+import { automationsQuery, runsQuery } from "@/lib/queries"
 import { ago } from "@/lib/time"
 import { useApiMutation } from "@/lib/use-api-mutation"
-import {
-  EVENT_KINDS,
-  runOutcome,
-  runStatusLabel,
-  SCHEDULE_PRESETS,
-  triggerLabel,
-} from "./automation-format"
+import { AutomationForm } from "./automation-form"
+import { runOutcome, runStatusLabel, triggerLabel } from "./automation-format"
 import { SettingsListSkeleton } from "./settings-list-skeleton"
 import { SettingsSection } from "./settings-section"
 
@@ -49,7 +35,9 @@ export function AutomationsSection() {
         </>
       }
     >
-      <NewAutomation onCreated={reload} />
+      <div className="rounded-lg border bg-card p-4">
+        <AutomationForm onCreated={reload} />
+      </div>
 
       {isPending ? (
         <SettingsListSkeleton />
@@ -81,134 +69,6 @@ export function AutomationsSection() {
 
       <Activity />
     </SettingsSection>
-  )
-}
-
-function NewAutomation({ onCreated }: { onCreated: () => void }) {
-  const { data: agents } = useQuery(agentsQuery())
-  const [agentId, setAgentId] = useState("")
-  const [instruction, setInstruction] = useState("")
-  const [kind, setKind] = useState<AutomationTrigger["kind"]>("manual")
-  const [cron, setCron] = useState<string>(SCHEDULE_PRESETS[0].cron)
-  const [on, setOn] = useState<string>(EVENT_KINDS[0].id)
-  const [route, setRoute] = useState<"auto" | "proposal">("proposal")
-  const tz = useMemo(() => Intl.DateTimeFormat().resolvedOptions().timeZone || "UTC", [])
-
-  const buildTrigger = (): AutomationTrigger =>
-    kind === "schedule" ? { kind, cron, tz } : kind === "event" ? { kind, on } : { kind }
-
-  const create = useApiMutation({
-    mutationFn: () =>
-      api.createAutomation({
-        agentId,
-        trigger: buildTrigger(),
-        instruction: instruction.trim(),
-        route,
-      }),
-    success: "Automation created",
-    onSuccess: () => {
-      setInstruction("")
-      onCreated()
-    },
-  })
-  const ready = agentId !== "" && instruction.trim() !== ""
-
-  return (
-    <div className="flex flex-col gap-3 rounded-lg border bg-card p-4">
-      <div className="flex flex-wrap gap-2">
-        <Select value={agentId} onValueChange={setAgentId}>
-          <SelectTrigger data-testid="automation-agent" aria-label="Agent" className="w-45">
-            <SelectValue placeholder="Pick an agent" />
-          </SelectTrigger>
-          <SelectContent>
-            {(agents ?? []).map((a) => (
-              <SelectItem key={a.id} value={a.id}>
-                @{a.name}
-              </SelectItem>
-            ))}
-          </SelectContent>
-        </Select>
-        <Select value={route} onValueChange={(v) => setRoute(v as "auto" | "proposal")}>
-          <SelectTrigger data-testid="automation-route" aria-label="Route" className="w-45">
-            <SelectValue />
-          </SelectTrigger>
-          <SelectContent>
-            <SelectItem value="proposal">Propose for review</SelectItem>
-            <SelectItem value="auto">Publish, then review</SelectItem>
-          </SelectContent>
-        </Select>
-      </div>
-
-      <Textarea
-        data-testid="automation-instruction"
-        aria-label="Instruction"
-        placeholder="What should the agent do? e.g. Keep this doc's dates and statuses current."
-        value={instruction}
-        onChange={(e) => setInstruction(e.target.value)}
-        rows={2}
-      />
-
-      <Tabs value={kind} onValueChange={(v) => setKind(v as AutomationTrigger["kind"])}>
-        <TabsList>
-          <TabsTrigger data-testid="automation-trigger-manual" value="manual">
-            Manual
-          </TabsTrigger>
-          <TabsTrigger data-testid="automation-trigger-schedule" value="schedule">
-            Schedule
-          </TabsTrigger>
-          <TabsTrigger data-testid="automation-trigger-event" value="event">
-            Event
-          </TabsTrigger>
-        </TabsList>
-      </Tabs>
-      {kind === "manual" && (
-        <p className="text-sm text-muted-foreground">Runs only when you press Run now.</p>
-      )}
-      {kind === "schedule" && (
-        <div className="flex flex-wrap items-center gap-2">
-          <Select value={cron} onValueChange={setCron}>
-            <SelectTrigger data-testid="automation-schedule" aria-label="Schedule" className="w-60">
-              <SelectValue />
-            </SelectTrigger>
-            <SelectContent>
-              {SCHEDULE_PRESETS.map((p) => (
-                <SelectItem key={p.id} value={p.cron}>
-                  {p.label}
-                </SelectItem>
-              ))}
-            </SelectContent>
-          </Select>
-          <span className="font-mono text-2xs text-muted-foreground">{tz}</span>
-        </div>
-      )}
-      {kind === "event" && (
-        <Select value={on} onValueChange={setOn}>
-          <SelectTrigger data-testid="automation-event" aria-label="Event" className="w-72">
-            <SelectValue />
-          </SelectTrigger>
-          <SelectContent>
-            {EVENT_KINDS.map((e) => (
-              <SelectItem key={e.id} value={e.id}>
-                {e.label}
-              </SelectItem>
-            ))}
-          </SelectContent>
-        </Select>
-      )}
-
-      <div className="flex justify-end">
-        <Button
-          data-testid="automation-create"
-          variant="secondary"
-          size="sm"
-          onClick={() => ready && create.mutate()}
-          loading={create.isPending}
-          disabled={create.isPending || !ready}
-        >
-          {create.isPending ? "Creating…" : "Create automation"}
-        </Button>
-      </div>
-    </div>
   )
 }
