+++ packages/db/src/pg.ts
@@ -107,7 +107,6 @@ import {
   inArray,
   isNotNull,
   isNull,
-  lt,
   lte,
   ne,
   notExists,
